+++ tests/Database/SelectFields/InterfaceTests/CommentType.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Rebing\GraphQL\Tests\Database\SelectFields\InterfaceTests;
 
-use GraphQL\Type\Definition\Type;
 use Rebing\GraphQL\Support\Facades\GraphQL;
 use Rebing\GraphQL\Support\Type as GraphQLType;
 use Rebing\GraphQL\Tests\Support\Models\Comment;
@@ -20,12 +19,7 @@ class CommentType extends GraphQLType
     {
         $interface = GraphQL::type('LikableInterface');
 
-        return [
-                'title' => [
-                    'type' => Type::nonNull(Type::string()),
-                    'alias' => 'title',
-                ],
-            ] + $interface->getFields();
+        return $interface->getFields();
     }
 
     public function interfaces(): array